+++ tests/plugins/test-woocommerce-memberships.php
@@ -1,4 +1,4 @@
-<?php // phpcs:ignoreFile
+<?php // phpcs:disable Squiz.Commenting, Universal.Files, Generic.Files
 /**
  * Class Newsletters Test Woocommerce Memberships.
  *
@@ -12,9 +12,11 @@ use Newspack\Newsletters\Subscription_Lists;
  * Test Woocommerce Memberships.
  */
 class Woocommerce_Memberships_Test extends \WP_UnitTestCase {
+	use WC_Memberships_Setup;
+
 	public static $subscription_list_id = 1;
 
-	public static function setup_membership_plans() {
+	public static function setup_test_memberships() {
 		$membership_plan_rule = new WC_Memberships_Membership_Plan_Rule(
 			[
 				'content_type_name' => Subscription_Lists::CPT,
@@ -28,20 +30,11 @@ class Woocommerce_Memberships_Test extends \WP_UnitTestCase {
 		return [ $membership_plan ];
 	}
 
-	public function test_is_subscription_list_tied_to_plan() { // phpcs:ignore Squiz.Commenting.FunctionComment.Missing
+	public function test_is_subscription_list_tied_to_plan() {
 		$subscription_list_tied_to_plan = Woocommerce_Memberships::is_subscription_list_tied_to_plan( self::$subscription_list_id );
 		$this->assertTrue( $subscription_list_tied_to_plan );
 
 		$subscription_list_not_tied_to_plan = Woocommerce_Memberships::is_subscription_list_tied_to_plan( 0 );
 		$this->assertFalse( $subscription_list_not_tied_to_plan );
 	}
-}
-
-/**
- * Mock wc_memberships_get_membership_plans.
- *
- * @return array
- */
-function wc_memberships_get_membership_plans() {
-	return Woocommerce_Memberships_Test::setup_membership_plans();
 }